add logic

+++ app/Http/Controllers/PrkMaterialController.php
@@ -3,7 +3,11 @@
 namespace App\Http\Controllers;
 
 use App\Http\Helpers\ResponseHelper;
-use App\Models\PrkMaterial;
+use App\Models\{
+    PrkMaterial,
+    Skki,
+    SkkiMaterial
+};
 use Illuminate\Http\Request;
 
 class PrkMaterialController extends Controller {
@@ -24,6 +28,7 @@ class PrkMaterialController extends Controller {
             'kode_normalisasi' => $request->kode_normalisasi,
             'nama_material' => $request->nama_material,
             'jumlah' => $request->jumlah,
+            'stok' => $request->stok,
             'harga' => $request->harga,
             'satuan' => $request->satuan,
             'prk_id' => $prk_id
@@ -40,6 +45,11 @@ class PrkMaterialController extends Controller {
 
     public function delete($prk_id, $material_id, Request $request) {
         $material = PrkMaterial::find($material_id);
+        $skki_material = SkkiMaterial::where(['kode_normalisasi' => $material->kode_normalisasi, 'prk_id' => $prk_id])->first();
+        if($skki_material) {
+            $skki = Skki::find($skki_material->skki_id);
+            return $this->response->bad_request('Material ini digunakan oleh SKKI #'.$skki->nomor_skki.' (ID: '.$skki->id.')');
+        }
         if($material->delete()) {
             return $this->response->success();
         }
